Volumetric functions and desinfection function
Added two functions for the pipes and the commercial tank volumetric
computation.
Added a desinfection function to compute the total quantities by element
for desinfection procedure (considering the number of elements and the
volume). 
It is missing a higher function that performs the sum of all quantities
for computing the total quantity of the system.