Backup uploader
The application is able to upload the backup file to the selected cloud
storage. The unique cloud storage uploader developed is for Dropbox but
another services can be supported extending the corresponding abstract
class.

NOTE: This function is only for this branch. Not will be joined to the
master branch soon.